Clean up makedirs().
makedirs() exist_ok=True.  Use it to avoid FileExistsError.  We
are now logging exceptions vi stderr in Daemon.py, so let exceptions
propagate up for debugging.

+++ grs/Kernel.py
@@ -96,7 +96,8 @@ class Kernel():
 
         cwd = os.getcwd()
         os.chdir(image_dir)
-        os.unlink(tarball_path)
+        if os.path.isfile(tarball_path):
+            os.unlink(tarball_path)
         cmd = 'tar -Jcf %s .' % tarball_path
         Execute(cmd, timeout=600, logfile=self.logfile)
         os.chdir(cwd)
